+++ lib/features/scriptInput/controller/script_input_controller.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
-import 'package:swm_peech_flutter/features/common/dataSource/local_script_stroage.dart';
+import 'package:swm_peech_flutter/features/common/dataSource/local_script_storage.dart';
 import 'package:swm_peech_flutter/features/scriptInput/dataSource/script_expected_time_data_source.dart';
 
 import '../../common/models/script_expected_time_model.dart';
