+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get_navigation/src/root/get_material_app.dart';
-import 'package:swm_peech_flutter/features/common/dataSource/local_practice_mode_stroage.dart';
-import 'package:swm_peech_flutter/features/common/dataSource/local_script_stroage.dart';
+import 'package:swm_peech_flutter/features/common/dataSource/local_practice_mode_storage.dart';
+import 'package:swm_peech_flutter/features/common/dataSource/local_script_storage.dart';
 import 'package:swm_peech_flutter/routers/routers.dart';
 
 void main() {
